+++ packages/core-json-rpc/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- Return the encoded WIF for BIP38 wallets instead of the encrypted WIF
+
 ## 0.2.0 - 2018-12-03
 
 ### Changed